다이얼로그에서 기계이름만 맵프래그먼트로 옮기게 변경 필요

+++ app/src/main/java/com/im_geokjeong/model/Office.kt
@@ -1,10 +1,7 @@
 package com.im_geokjeong.model
 
-import android.os.Parcel
-import android.os.Parcelable
-
 data class Office(
-    val id: Int,
+    val id: Long,
     val name: String,
     val phoneNumber: String,
     val streetNameAddress: String,
