+++ app/src/main/java/com/im_geokjeong/repository/officedetail/OfficeDetailDataSource.kt
@@ -4,5 +4,5 @@ import com.im_geokjeong.model.OfficeDetailResponse
 
 interface OfficeDetailDataSource {
 
-    suspend fun getOfficeDetail(officeId: Int): OfficeDetailResponse
+    suspend fun getOfficeDetail(officeId: Long): OfficeDetailResponse
 }
